Add new recipe
Added new Recipe for the website. I used the unordered list for the Ingredients as this does not need to be ordered for this recipe. I used the ordered list to list the steps involved in making the coquito. This makes it easier to follow as the ordered list is numbered.